requirements.txt
fastapi==0.88.0
pydantic==1.10.2
requests==2.28.1
uvicorn==0.20.0
python-multipart==0.0.5
fastapi-jwt-auth==0.5.0
python-dotenv==0.21.0
pymongo==4.3.3